requirements.txt
flask
fastapi
uvicorn
ezdxf
PyMuPDF
pdfplumber
ultralytics
opencv-python
pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
# tensorflow  # Commented out - not available for Python 3.13 on Windows
torch
torchvision
pytesseract
easyocr
shapely
scipy
networkx
plotly
python-multipart
python-dotenv
requests
jupyter
notebook
ipywidgets
tqdm
pyyaml
click
watchdog
psutil
py7zr
rarfile
# ONNX packages for model export
onnx>=1.18.0
onnxruntime>=1.22.0
onnxslim>=0.1.58
